Changes in setup.py and build.sh to dynamically determine python interpreter location

+++ setup.py
@@ -9,18 +9,22 @@ else:
     with open("README.md", "r") as fh:
         long_description = fh.read()
         home = os.path.expanduser("~")
-        py_loc = f'#!{sys.executable}\n'
-        dest_file = f"{home}/.local/bin/watiba-c"
-        with open("watiba/version.py") as f:
-            ver = f.read()
-            ver_line = f'versions = ["Watiba {ver}", "Python 3.8"]\n'
-            with open(dest_file, 'w') as wf:
-                wf.write(py_loc)
-                wf.write(ver_line)
-                with open("watiba/watiba-c.py", 'r') as rf:
-                    wf.write(rf.read())
+        if os.path.exists(f'{home}/.local/bin'):
+            py_loc = f'#!{sys.executable}\n'
+            dest_file = f"{home}/.local/bin/watiba-c"
+            with open("watiba/version.py") as f:
+                ver = f.read()
+                ver_line = f'versions = ["Watiba {ver}", "Python 3.8"]\n'
+                with open(dest_file, 'w') as wf:
+                    wf.write(py_loc)
+                    wf.write(ver_line)
+                    with open("watiba/watiba-c.py", 'r') as rf:
+                        wf.write(rf.read())
 
-            os.chmod(dest_file, 0o0766)
+                os.chmod(dest_file, 0o0766)
+        else:
+            print(f"WARNING: {home}/.local/bin not found on this system.  "
+                  f"Please copy watiba/watiba-c-bin.py to a location within your PATH")
 
 
 with open("watiba/version.py", "r") as fh:
